refactor to db error extractor
