+++ packages/nocodb/tests/unit/index.test.ts
@@ -5,6 +5,7 @@ import modelTests from './model/index.test';
 import TestDbMngr from './TestDbMngr';
 import { dataApiV3Test } from './rest/tests/dataApiV3/index.test';
 import { processorTests } from './processor/index.test';
+import { errorTests } from './error/index.test';
 
 process.env.NODE_ENV = 'test';
 process.env.TEST = 'true';
@@ -20,6 +21,8 @@ dotenv.config({
   await TestDbMngr.init();
 
   modelTests();
+  formulaTests();
+  errorTests();
   restTests();
   processorTests();
   await dataApiV3Test();
